Start moving upgrades to an apt-repository-based system.

+++ package/fa_adept_client.tcl
@@ -589,47 +589,70 @@ set caDir [file join [file dirname [info script]] "ca"]
 		logger "performing $type update, action: $row(action)"
 
 		set restartPiaware 0
+		set updatedPackageLists 0
+		set ok 1
 		foreach action [split $row(action) " "] {
-			switch $action {
-				"full" {
-					update_operating_system_and_packages
+			if {$action in {full packages piaware dump1090}} {
+				# these actions require that the package lists
+				# are up to date, but we don't want to do it
+				# several times
+				if {!$updatedPackageLists} {
+					set ok [update_package_lists]
 				}
+			}
 
-				"packages" {
-					upgrade_raspbian_packages
+			if {!$ok} {
+				logger "skipping action $action"
+				continue
+			}
+
+			switch $action {
+				"full" - "packages" {
+					set ok [upgrade_all_packages]
 				}
 
 				"piaware" {
 					# only restart piaware if upgrade_piaware said it upgraded
 					# successfully
-					set restartPiaware [upgrade_piaware]
+					set ok [upgrade_piaware]
+					if {$ok} {
+						set restartPiaware 1
+					}
 				}
 
 				"restart_piaware" {
 					set restartPiaware 1
+					set ok 1
 				}
 
 				"dump1090" {
 					# try to upgrade dump1090 and if successful, restart it
-					if {[upgrade_dump1090]} {
+					set ok [upgrade_dump1090]
+					if {$ok} {
 						attempt_dump1090_restart
 					}
 				}
 
 				"restart_dump1090" {
 					attempt_dump1090_restart
+					set ok 1
 				}
 
 				"reboot" {
 					reboot
+					# don't run anything further
+					set ok 0
 				}
 
 				"halt" {
 					halt
+					# don't run anything further
+					set ok 0
 				}
 
 				default {
 					logger "unrecognized update action '$action', ignoring..."
+					set ok 0
 				}
 			}
 		}
